Fixed currency rate, added currency saving
Currency is now saved in local storage (whether on an Android phone or a
PC/Mac). If you exit the application and come back to it at a later
time, your most recent currency amount will be retrieved. Only when the
application is uninstalled will the currency storage file be removed.